+++ idmap/apps.py
@@ -3,6 +3,7 @@ from django.apps import AppConfig
 
 class IdMapConfig(AppConfig):
     name = 'idmap'
+    verbose_name = 'Django identity mapper'
 
     def ready(self):