+++ idmap/signals.py
@@ -1,21 +1,31 @@
+"""
+Signals are used to automatically flush the idmap cache on finish request,
+migrate / syncdb and instance deletion, and to make sure to catch cascades
+"""
+
+from django.dispatch import Signal, receiver
 from django.core.signals import request_finished
 from django.db.models.signals import pre_delete, post_migrate
 
 
-# Flush cache after syncdb
-def flush_cache(**kwargs):
-    from .models import SharedMemoryModel
-    for model in SharedMemoryModel.__subclasses__():
-        model.flush_instance_cache(flush_sub=True)
-
-request_finished.connect(flush_cache)
-post_migrate.connect(flush_cache)
+pre_flush = Signal(providing_args=['using'])
+post_flush = Signal(providing_args=['using'])
 
 
-# Remove instance from cache upon deletion
+@receiver((post_migrate, request_finished))
+def flush_idmap(using=None, **kwargs):
+    """
+    Flushes the idmap cache on migrate and on request end
+    """
+    from .functions import flush
+    flush(db=using)
+
+
+@receiver(pre_delete)
 def flush_cached_instance(sender, instance, **kwargs):
-    # XXX: Is this the best way to make sure we can flush?
-    if not hasattr(instance, 'flush_cached_instance'):
-        return
-    sender.flush_cached_instance(instance)
-pre_delete.connect(flush_cached_instance)
+    """
+    Flushes a deleted instance from the idmap cache
+    """
+    from .models import IdMapModel
+    if issubclass(sender, IdMapModel):
+        sender.flush_cached_instance(instance)