+++ tests/subclasses.py
@@ -1,14 +1,15 @@
-from django.test import TestCase
 from django.utils import six
 
+from idmap import flush
+
 from .app.models import Article, SubArticle, Category, RegularCategory
-from idmap import flush_cache
+
+from ._base import TestCase
 
 
 class SubclassesTests(TestCase):
 
     def setUp(self):
-        flush_cache()
         category = Category.objects.create(name="Category")
         regcategory = RegularCategory.objects.create(name="RegCategory")
 
@@ -20,7 +21,7 @@ class SubclassesTests(TestCase):
                                       category=category,
                                       category2=regcategory)
 
-    def testFlushSubArticle(self):
+    def test_flush_sub_article(self):
 
         # make a list of Articles so that they're in cache
         article_list = list(Article.objects.all())
@@ -40,14 +41,14 @@ class SubclassesTests(TestCase):
         for pk, __ in sub_pkids:
             self.assertIsNone(SubArticle.get_cached_instance(pk))
 
-    def testFlushAll(self):
+    def test_flush_all(self):
 
         # make a list of Articles and SubArticles so that they're in cache
         list(Article.objects.all())
         sub_pks = [sa.pk for sa in SubArticle.objects.all()]
 
         # should flush Article and SubArticle's caches
-        flush_cache()
+        flush()
 
         for pk in sub_pks:
             self.assertIsNone(SubArticle.get_cached_instance(pk))