+++ CHANGES.rst
@@ -2,6 +2,18 @@ django-idmap - changes
 ======================
 
 
+v1.0 (22-05-2017)
+-----------------
+
+Added:
+- Support of multiple databases for one model
+- Renames ``flush_cache`` function to ``flush``
+- ``pre_flush`` and ``post_flush`` signals
+- Support for Django 1.11 and Python 3.6
+- Support of proxy models using proxy groups
+- Support of options in django's Meta
+
+
 v0.4 (09-12-2015)
 -----------------